requirements.txt
Pillow>=9.0.0
tqdm>=4.62.0
numpy>=1.20.0